Añadiendo más lógica de HU reset-password
