--- src/main/java/com/iRead/backendproyect/services/ReserPasswordService.java
@@ -1,11 +1,0 @@
-package com.iRead.backendproyect.services;
-
-import lombok.AllArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@AllArgsConstructor
-public class ReserPasswordService {
-
-
-}
